Fix Deployment of Flagger loadtester to include the correct label
The Deployment of the Flagger loadtester did not contain the correct
label app.kubernetes.io/name. This label is used for the Flagger
deployment and it is also used in the PodDisruptionBudget for
the Flagger operator. I added the same label to the Flagger
load tester to make the PodDisruptionBudget work correctly
for the Flagger loadtester.

Signed-off-by: Marcus Rodan <marcusrodan@gmail.com>