Rework selector labels
Addresses #1312, #1257, #1115

Retains all `matchLabels` in the primary `Deployment` or `DaemonSet`, instead of picking the first one that matches the `-selector-labels`.

Recreates the primary `Deployment` or `DaemonSet` if the `matchLabels` of the existing object are not equal. This must be done since `matchLabels` is an immutable field. The recreate happens before promoting the canary to primary (i.e. no traffic is being routed to the primary).

Still appends `-primary` to the value of the first label that matches `-selector-labels`. This ensures that the canary and primary have a unique set of selector labels.

`Service` selector also uses the full set of `matchLabels` and overrides the value of the first matching `-selector-labels` to route to canary or primary.

Signed-off-by: Kevin Snyder <kevin.snyder@gusto.com>